code re factoring done
