+++ server/tests/seeds/visitorData.js
@@ -1,0 +1,69 @@
+const notification = {
+  title: 'Voluptatum ',
+  description: 'Neque ',
+  file: 'http:',
+  tags: ['tag1'],
+  link: 'https',
+};
+
+const questionPaper = {
+  subject: 'maths',
+  year: '2018',
+  semester: '4th sem',
+  branch: 'it',
+  link: 'httpskjdf',
+};
+
+const event = {
+  name: 'shubham birthday',
+  place: 'nlksdjsdfn',
+  description:
+    ' Lorem Ipsum is simply dummy text of the printing and Ipsum is simply dummy text of the printing and Ipsum is simply dummy text of the printing and Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry ',
+};
+
+const syllabus = {
+  branch: 'it branch ',
+  codeNo: '20150',
+  semester: '4',
+  subject: 'maths',
+
+  l: '2524',
+  tp: '422',
+  credits: '24524',
+  status: 'dskfjfkd',
+  period: 'fgsdhgfdsgh',
+  file: 'fdhfdshdsh',
+};
+
+const timetable = {
+  branch: 'it',
+  semester: 'gfhfghfg',
+  title: 'hgfgdhfhf',
+  file: 'dgfdgg',
+};
+
+const student = {
+  name: 'Mabel',
+  rollNo: 4202070311602303,
+  location: 'Port Constantinview',
+  dateOfBirth: 'Mon Jun 26 1995 02:27:14 GMT+0530 (India Standard Time)',
+  gender: 'male',
+  branch: 'IT',
+  email: 'hello@gmail.com',
+  admittedIn: 'Mon Sep 11 2016 01:42:00 GMT+0530 (India Standard Time)',
+  bio:
+    'Ducimus eius est ea. Dolorem adipisci aliquid nulla autem nihil voluptas. Quis saepe numquam possimus qui et ut laboriosam est. Iste eaque omnis velit nesciunt deserunt qui.',
+  auth: {
+    provider: 'Eaque assumenda molestiae quod aut saepe et.',
+    providerId: 'voluptas',
+  },
+};
+
+module.exports = {
+  notification,
+  questionPaper,
+  event,
+  syllabus,
+  timetable,
+  student,
+};
