+++ server/utils/urls.js
@@ -1,29 +1,29 @@
 // ------- visitiors routes urls ---------
 
-const getnotifications = '/s/visitor/getnotifications';
-const getallnotifications = '/s/visitor/getallnotifications';
-const getallQuestionPaper = '/s/visitor/getallQuestionPaper';
-const getevents = '/s/visitor/getevents';
-const getallevents = '/s/visitor/getallevents';
-const getsyll = '/s/visitor/getsyll';
-const gettt = '/s/visitor/gettt';
-const getteacher = '/s/visitor/getteacher';
-const getstudent = '/s/visitor/getstudent';
-const searchstudentbyname = '/s/visitor/searchstudentbyname';
-const searchstudentbyrollno = '/s/visitor/searchstudentbyrollno';
-const verifyemail = '/s/visitor/verifyemail';
+const GET_NOTIFICATIONS = '/s/visitor/getnotifications';
+const GET_ALL_NOTIFICATIONS = '/s/visitor/getallnotifications';
+const GET_ALL_QUESTION_PAPAER = '/s/visitor/getallQuestionPaper';
+const GET_EVENTS = '/s/visitor/getevents';
+const GET_ALL_EVENTS = '/s/visitor/getallevents';
+const GET_SYLLABUS = '/s/visitor/getsyll';
+const GET_TIMETABEL = '/s/visitor/gettt';
+const GET_TEACHER = '/s/visitor/getteacher';
+const GET_STUDENT = '/s/visitor/getstudent';
+const SEARCH_STUDENT_BY_NAME = '/s/visitor/searchstudentbyname';
+const SEARCH_STUDENT_BY_ROLLNO = '/s/visitor/searchstudentbyrollno';
+const VERIFY_EMAIL = '/s/visitor/verifyemail';
 
 module.exports = {
-  getnotifications,
-  getallnotifications,
-  getallQuestionPaper,
-  getevents,
-  getallevents,
-  getsyll,
-  gettt,
-  getteacher,
-  getstudent,
-  searchstudentbyname,
-  searchstudentbyrollno,
-  verifyemail,
+  GET_NOTIFICATIONS,
+  GET_ALL_NOTIFICATIONS,
+  GET_ALL_QUESTION_PAPAER,
+  GET_EVENTS,
+  GET_ALL_EVENTS,
+  GET_SYLLABUS,
+  GET_TIMETABEL,
+  GET_TEACHER,
+  GET_STUDENT,
+  SEARCH_STUDENT_BY_NAME,
+  SEARCH_STUDENT_BY_ROLLNO,
+  VERIFY_EMAIL,
 };
